a more general solution in IsolatedProcess
