+++ bluecellulab/analysis/inject_sequence.py
@@ -8,9 +8,9 @@ import numpy as np
 from bluecellulab.cell.core import Cell
 from bluecellulab.cell.template import TemplateParams
 from bluecellulab.simulation.neuron_globals import NeuronGlobals
+from bluecellulab.simulation.parallel import IsolatedProcess
 from bluecellulab.simulation.simulation import Simulation
 from bluecellulab.stimulus.factory import Stimulus, StimulusFactory
-from bluecellulab.utils import IsolatedProcess
 
 
 class StimulusName(Enum):
